Use bus topic manager in publisher
