+++ PubSub.Application/Publish/IPublisher.cs
@@ -1,0 +1,7 @@
+﻿namespace PubSub.Application.Publish
+{
+    public interface IPublisher
+    {
+        Task Publish(Publication publication);
+    }
+}
