+++ PubSub.Application/Publish/Publication.cs
@@ -1,0 +1,4 @@
+﻿namespace PubSub.Application.Publish
+{
+    public record Publication(string Queue, string Message);
+}
